# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "operationId": "UserController_getProfileById",
        "summary": "Get user info by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "summary": "Update User",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/api/user/profile": {
      "get": {
        "operationId": "UserPermissionController_user__GetMyProfile",
        "summary": "Get user profile by JWT",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "UserPermissionController_user__UpdateMyProfile",
        "summary": "Get user profile by JWT",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/class-joined": {
      "get": {
        "operationId": "UserPermissionController_getClassJoined",
        "summary": "Get class joined",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/authorization/{id}": {
      "put": {
        "operationId": "UserPermissionController_user__AuthorizationTeacher",
        "summary": "Authorization teacher by JWT",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/change-password": {
      "put": {
        "operationId": "UserPermissionController_user__ChangePassword",
        "summary": "Change password by JWT",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response data template after processing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The response data template after processing encounters exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultResponse"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bluequant API APIs document",
    "description": "",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ErrorDataResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "required"
          },
          "message": {
            "type": "string",
            "description": "required"
          },
          "error": {
            "type": "string",
            "description": "required"
          }
        },
        "required": [
          "code",
          "message",
          "error"
        ]
      },
      "DefaultResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "required"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDataResponse"
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "updatedAt": {
            "type": "date",
            "description": "Example: 2025-10-18T02:17:46.007Z",
            "default": "2025-10-18T02:17:46.007Z"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarLocation": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "birthDay": {
            "type": "datetime"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ],
            "default": "MALE"
          },
          "code": {
            "type": "string"
          },
          "isDeleted": {
            "type": "bit",
            "default": 0
          },
          "isApproved": {
            "type": "bit",
            "default": 0
          },
          "schoolId": {
            "type": "bigint"
          },
          "houseStreet": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        },
        "required": [
          "createdDate",
          "updatedAt",
          "password",
          "name",
          "avatarLocation",
          "email",
          "phoneNumber",
          "address",
          "birthDay",
          "gender",
          "code",
          "isDeleted",
          "isApproved",
          "schoolId",
          "houseStreet",
          "ward",
          "district",
          "city"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "required",
            "default": "John Doe",
            "maxLength": 50
          },
          "address": {
            "type": "string",
            "description": "not required",
            "maxLength": 255
          },
          "classRoomName": {
            "type": "string",
            "description": "required",
            "maxLength": 50
          },
          "schoolName": {
            "type": "string",
            "description": "required",
            "maxLength": 50
          },
          "birthDay": {
            "type": "string",
            "description": "not required"
          }
        },
        "required": [
          "name",
          "classRoomName",
          "schoolName"
        ]
      },
      "UpdateUserProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "not required",
            "default": "thanhtung",
            "maxLength": 50
          },
          "avatarLocation": {
            "type": "string",
            "description": "not required"
          },
          "phoneNumber": {
            "type": "string",
            "description": "not required"
          },
          "birthDay": {
            "type": "string",
            "description": "not required"
          },
          "address": {
            "type": "string",
            "description": "not required"
          },
          "gender": {
            "type": "string",
            "description": "not required",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "schoolId": {
            "type": "string",
            "description": "not required"
          },
          "houseStreet": {
            "type": "string",
            "description": "not required"
          },
          "ward": {
            "type": "string",
            "description": "not required"
          },
          "district": {
            "type": "string",
            "description": "not required"
          },
          "city": {
            "type": "string",
            "description": "not required"
          }
        }
      },
      "ChangeUserPasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "not required",
            "default": "123456"
          },
          "newPassword": {
            "type": "string",
            "description": "not required",
            "default": "666666"
          },
          "confirmPassword": {
            "type": "string",
            "description": "not required",
            "default": "888888"
          }
        }
      },
...
    }
  }
}